refactor(web): isolate GenUI feature boundary
