+++ examples/web/src/entries/genui.js
@@ -1,0 +1,1 @@
+import '../features/genui/browser/mount.js';
