+++ examples/web/src/features/genui/browser/styles.css
@@ -1,0 +1,47 @@
+    @keyframes fadeIn {
+      from { opacity: 0; transform: translateY(-4px); }
+      to { opacity: 1; transform: translateY(0); }
+    }
+
+    .btn {
+      padding: 5px 10px; background: #3c3c3c; color: #abb2bf;
+      border: 1px solid #555; border-radius: 3px; cursor: pointer;
+      font-size: 11px; font-family: inherit; transition: background 0.15s;
+    }
+    .btn:hover { background: #4c4c4c; border-color: #666; }
+    .btn-primary {
+      padding: 5px 10px; background: #4d78cc; color: #fff;
+      border: 1px solid #4d78cc; border-radius: 3px; cursor: pointer;
+      font-size: 11px; font-family: inherit; transition: background 0.15s;
+    }
+    .btn-primary:hover { background: #5c8ae6; }
+    .btn-primary:disabled { opacity: 0.4; cursor: not-allowed; }
+    .view-tab {
+      padding: 4px 10px; background: #2d2d2d; border: 1px solid #2c313a;
+      border-radius: 3px; cursor: pointer; font-size: 11px; color: #565c64;
+      font-family: inherit;
+    }
+    .view-tab.active { background: #3c3c3c; color: #abb2bf; border-color: #4ec9b0; }
+    .view-tab:hover { background: #3c3c3c; }
+    .view-panel { display: none; }
+    .view-panel.active { display: flex; }
+    .node-id { display: inline-block; min-width: 28px; padding: 0 3px; border-radius: 2px; text-align: center; font-weight: bold; font-size: 10px; }
+    .node-id.stable { background: rgba(78, 201, 176, 0.25); color: #4ec9b0; }
+    .node-id.new { background: rgba(240, 198, 116, 0.25); color: #f0c674; }
+    .tree-toggle { color: #565c64; margin-right: 3px; }
+    .tree-node { padding-left: 16px; border-left: 1px solid #2c313a; margin: 1px 0; }
+    .tree-node-header { cursor: pointer; user-select: none; }
+    .tree-node-header:hover { background: rgba(255,255,255,0.02); }
+    .error-item { color: #f48771; padding: 2px 0; font-size: 11px; border-left: 2px solid #f48771; padding-left: 6px; margin-bottom: 2px; }
+    .binding-badge { color: #82aaff; background: rgba(130,170,255,0.12); border: 1px solid rgba(130,170,255,0.3); border-radius: 3px; padding: 2px 6px; }
+    .summary-card, .detail-field { padding: 8px 10px; background: rgba(30,34,42,0.7); border: 1px solid #2c313a; border-radius: 4px; }
+    .summary-card dt, .detail-field dt { color: #565c64; font-size: 10px; text-transform: uppercase; letter-spacing: 0.04em; }
+    .summary-card dd, .detail-field dd { color: #abb2bf; font-size: 12px; margin-top: 3px; }
+    .order-row { border-bottom: 1px solid rgba(44,49,58,0.7); cursor: pointer; }
+    .order-row:hover { background: rgba(78,201,176,0.06); }
+    .order-row.selected { background: rgba(78,201,176,0.14); box-shadow: inset 2px 0 #4ec9b0; }
+    .status-chip { display: inline-block; border-radius: 999px; padding: 2px 7px; font-size: 10px; }
+    .status-paid { color: #9fe2c9; background: rgba(78,201,176,0.14); }
+    .status-pending { color: #f0c674; background: rgba(240,198,116,0.14); }
+    .status-review { color: #82aaff; background: rgba(130,170,255,0.14); }
+    .feasibility-case.active { color: #e8e8f0; border-color: #8250df; background: rgba(130,80,223,0.18); }
