+++ examples/web/src/features/genui/core/data.ts
