+++ examples/web/src/features/genui/core/fixtures/incidents-csv.js
