+++ examples/web/src/features/genui/core/fixtures/inventory.json
