+++ examples/web/src/features/genui/core/genui-candidate-schema.js
