+++ examples/web/src/features/genui/core/genui-feasibility-demo.test.mjs
