+++ examples/web/src/features/genui/core/genui-feasibility-fixtures.js
