+++ examples/web/src/features/genui/core/genui-feasibility-fixtures.test.mjs
