+++ examples/web/src/features/genui/core/genui-feasibility-flow.js
