+++ examples/web/src/features/genui/core/genui-recorded-candidates.js
