+++ examples/web/src/features/genui/core/genui-spike-case.js
