+++ examples/web/src/features/genui/core/genui-spike-recipe.test.mjs
